Major fixes in all 3 tracks
Common:
* Made objects.py into a package because it got too big
* Added some utilities to all the tracks
* Made all code files unicode
* Reworked the way evaluators return their results
* Added silent comparison for testing purposes
* Standard and test markup files can now be loaded from the same
directory

Task 1:
* Expanded parent entity finding logic
* Fixed 'sibling' entity issues (several entities on a single span
* Reworked reports so that ignored objects/pairs are marked as such
* Entity skipping now works as intended

Task 2:
(no task-specific changes AFAIR)

Task 3:
* Entirely reworked the evaluator output
* IsPartOf is now processed properly
* Fixed entity name options search
* Reworked restrictions on the way facts are matched
