+++ scripts/dialent/task3/eval.py
@@ -1,3 +1,4 @@
+﻿# This module contains task 3 evaluation logic
 
 import os
 
@@ -6,7 +7,7 @@ from dialent.config import Tables
 from dialent.standard import Standard
 from dialent.task3.test import Test
 
-from dialent.objects import Fact
+from dialent.objects.fact import Fact
 from dialent.common.metrics import Metrics
 
 from dialent.task3.util import loadAllStandard
@@ -17,11 +18,15 @@ from dialent.task3.util import loadAllTest
 class Evaluator:
     """Evaluates the task 3 responses"""
 
+    # Tags used in statistics (everything but the ignored 'IsPartOf')
+    stat_tags = ['ownership', 'occupation', 'meeting', 'deal', 'overall']
+
     def __init__(self, hard_mode=False):
         self.hard_mode = hard_mode
 
-    def evaluate(self, std_path, test_path, output_path):
-        print('Running evaluation, this might take a while...')
+    def evaluate(self, std_path, test_path, output_path, is_silent=False):
+        if not is_silent:
+            print('Running evaluation, this might take a while...')
         std = loadAllStandard(std_path)
         test = loadAllTest(test_path)
         
@@ -29,7 +34,7 @@ class Evaluator:
             set([y.name for y in test]))
 
         assert(len(diff) == 0)
-        res = Metrics()
+        res = dict((x, Metrics()) for x in Evaluator.stat_tags)
 
         for i, s in enumerate(std):
             if not s.has_facts:
@@ -38,16 +43,36 @@ class Evaluator:
                 continue
             metrics = self.evaluateDocument(s, test[i])
             self.printReport(s.name, output_path)
-            res.add(metrics)
+            for tag in Evaluator.stat_tags:
+                res[tag].add(metrics[tag])
             
-        print(Metrics.header())
-        print(res.toLine())
+        if not is_silent:
+            print('TAG             ' + Metrics.header())
+            for tag in Evaluator.stat_tags:
+                print('{:15} '.format(tag) + res[tag].toLine())
+
+        return res
 
     def evaluateDocument(self, std, test):
-        self.optimizer = Optimizer(std, test, self.hard_mode)
-        self.optimizer.findSolution()
+        self.metrics = dict((x, Metrics()) for x in Evaluator.stat_tags)
+        self.clusters = []
+        for tag in Evaluator.stat_tags:
+            if tag == 'overall':
+                continue
+            tag_std = [s for s in std.facts if s.tag == tag]
+            tag_test = [t for t in test.facts if t.tag == tag]
+            self.optimizer = Optimizer(tag_std, tag_test, self.hard_mode)
+            self.optimizer.findSolution()
+            
+            self.metrics[tag].add(self.optimizer.metrics)
+            self.metrics['overall'].add(self.optimizer.metrics)
+            self.clusters.extend(self.optimizer.clusters)
         
-        return self.optimizer.metrics
+        return self.metrics
+
+    def buildReport(self):
+        """Build an evaluation report"""
+        return self.optimizer.describeMatching(self.clusters, self.metrics)
 
     def printReport(self, name, out_dir):
         """Print a detailed report on the document evaluation"""
@@ -56,21 +81,21 @@ class Evaluator:
 
         os.makedirs(out_dir, exist_ok=True)
         with open(os.path.join(out_dir, name + '.report.txt'), 'w', encoding='utf-8') as f:
-            f.write(self.optimizer.describeMatching())
+            f.write(self.buildReport())
 
 class Optimizer:
     """Optimizes the matching"""
 
     def __init__(self, std, test, hard_mode):
-        self.test = test.facts
+        self.test = test
         self.hard_mode = hard_mode
 
         if hard_mode:
             # hard mode, remove all facts with modality other than 'actual'
-            self.std = [x for x in std.facts if not x.has_easymode_modality]
+            self.std = [x for x in std if not x.has_easymode_modality]
         else:
             # easy mode, ignore all facts marked as difficult, and remove phase argument
-            self.std = std.facts
+            self.std = std
             for fact in self.std:
                 if fact.has_hardmode_difficulty:
                     fact.is_ignored = True
@@ -156,14 +181,17 @@ class Optimizer:
 
         return Metrics.create(tp_std, tp_test, n_std, n_test)
 
-    def describeMatching(self):
+    def describeMatching(self, clusters, metrics):
         """Returns a string description of the matching this optimizer built"""
-        res = '{:4}\t{:4}\t{:4}\t{:8}\n'.format('Res', 'Q_A', 'Q_Id', 'Facts')
-        res += '\n'.join([c.toInlineString() for c in self.clusters])
+        res = ''
+        for i, c in enumerate(clusters):
+            res += '---- #{} ----\n'.format(i+1)
+            res += c.toInlineString() + '\n'
         res += '\n\n'
         res += '-------METRICS------\n'
-        res += Metrics.header() + '\n'
-        res += self.metrics.toLine()
+        res += 'TAG             ' + Metrics.header() + '\n'
+        for tag in Evaluator.stat_tags:
+            res += '{:15} '.format(tag) + metrics[tag].toLine() + '\n'
         return res
 
 
@@ -187,16 +215,7 @@ class Cluster:
         
     def calculateQuality(self):
         """Calculate quality"""
-
-        if self.std == None or len(self.test) == 0 or self.std.is_ignored:
-            self.arg_quality = 0
-            self.id_quality = 0
-            self.quality = 0
-            return
-
-        # we're dealing with a proper matching
         self._doCalculateQuality()
-
         self.quality = (self.arg_quality + self.id_quality * self.arg_quality) / 2.0
 
     def _doCalculateQuality(self):
@@ -205,31 +224,41 @@ class Cluster:
         for t in self.test:
             t_args.extend(t.arguments)
 
+        s_args = self.std.arguments if self.std != None else []
+
         matched_t_args = set()
-        unmatched_t_args = set(t_args)
+        self.unmatched_t_args = set(t_args)
         matched_s_args = set()
-        unmatched_s_args = set(self.std.arguments)
-        t_by_s = dict([(s, []) for s in self.std.arguments])
-        s_by_t = dict()
+        self.unmatched_s_args = set(s_args)
+        self.t_by_s = dict([(s, []) for s in s_args])
+        self.s_by_t = dict()
+        
+        if self.std == None or len(self.test) == 0 or self.std.is_ignored:
+            # matching with objects on one side
+            self.arg_quality = 0
+            self.id_quality = 0
+            self.quality = 0
+            return
+
         for t in t_args:
             found_match = False
             for s in self.std.arguments:
                 if s.canMatch(t):
                     found_match = True
-                    if s in unmatched_s_args:
-                        unmatched_s_args.remove(s)
-                    unmatched_t_args.remove(t)
+                    if s in self.unmatched_s_args:
+                        self.unmatched_s_args.remove(s)
+                    self.unmatched_t_args.remove(t)
                     matched_s_args.add(s)
                     matched_t_args.add(t)
-                    t_by_s[s].append(t)
-                    s_by_t[t] = s
+                    self.t_by_s[s].append(t)
+                    self.s_by_t[t] = s
                     break
             if found_match:
                 continue
 
         n = sum(Tables.getArgumentWeight(x.name) for x in matched_s_args)
-        fp = sum(Tables.getArgumentWeight(x.name) for x in unmatched_t_args)
-        fn = sum(Tables.getArgumentWeight(x.name) for x in unmatched_s_args)
+        fp = sum(Tables.getArgumentWeight(x.name) for x in self.unmatched_t_args)
+        fn = sum(Tables.getArgumentWeight(x.name) for x in self.unmatched_s_args)
         self.arg_quality = n / float(n + fp + fn)
         if(self.arg_quality > 1):
             print(self)
@@ -241,8 +270,8 @@ class Cluster:
                 x = self.std.arguments[i]
                 y = self.std.arguments[j]
 
-                for t1 in t_by_s[x]:
-                    for t2 in t_by_s[y]:
+                for t1 in self.t_by_s[x]:
+                    for t2 in self.t_by_s[y]:
                         if t1.fact == t2.fact:
                             edges.add((x,y))
 
@@ -255,17 +284,47 @@ class Cluster:
     def toInlineString(self):
         """Create an inline description of a cluster"""
         if self.std != None and self.std.is_ignored:
-            res = '\tIGNORED\t\t'
+            res = 'Quality:\n\t(IGNORED)\n\n'
         else:
-            res = '{:.2f}\t{:.2f}\t{:.2f}\t'.format(self.quality, self.arg_quality, self.id_quality)
-        res += self.std.toInlineString() if self.std != None else '[---unmatched---]'
-        res += '\t=\t'
+            res = 'Quality:\n'
+            res += '\tArgument extraction quality = {:.2f}\n'.format(self.arg_quality)
+            res += '\tIdentification quality = {:.2f}\n'.format(self.id_quality)
+            res += '\tOVERALL = {:.2f}\n'.format(self.quality)
+            res += '\n'
+
+        res += 'STANDARD:\n\t' + (
+            self.std.toInlineString() if self.std != None else '[---unmatched---]')
+        
+        res += '\nTEST:\n\t'
         if len(self.test) == 0:
-            res += '[---unmatched---]'
+            res += '[---unmatched---]\n\n'
             return res
 
-        res += ', '.join([x.toInlineString() for x in self.test])
+        res += ', '.join([x.toInlineString() for x in self.test]) + '\n'
+
+        if self.std == None:
+            return res
+
+        res += '\nARGUMENTS:\n'
+        for arg in self.t_by_s:
+            if len(self.t_by_s[arg]) == 0:
+                assert(arg in self.unmatched_s_args)
+                continue
+
+            res += ('\t{} = {}\n'.format(
+                str(arg),
+                ', '.join(str(t) for t in self.t_by_s[arg])))
+
+        for arg in self.unmatched_s_args:
+            res += '\t{} = {}\n'.format(arg, '[---unmatched---]')
+
+        for arg in self.unmatched_t_args:
+            res += '\t{} = {}\n'.format('[---unmatched---]', arg)
+
         return res
+
+    def __str__(self):
+        return 'Cluster:\nS={}\nT={}\n'.format(self.std, self.test)
 
     @classmethod
     def unpairedTest(cls, test):
