+++ scripts/t1_eval.py
@@ -1,4 +1,4 @@
-# Runs the evaluation of the task 1 response
+﻿# Runs the evaluation of the task 1 response
 # Requires python 3 and numpy
 
 # Usage:
